s6-modules/requirements.txt
openai>=1.0.0
streamlit>=1.24.0
instructor>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0 